requirements.txt
python-telegram-bot==13.15
Flask==3.0.0
httpx~=0.24.0